Issue 20: Handle multipart uploads (#23)
This PR contains the following changes:

- Added the necessary logic to handle parallel multipart uploads and convert them into regular puts.
- Reworked the PUT and GET paths to make them async and be able to process and transfer data in streaming fashion.
- Added an end-to-end test that performs a multipart upload, downloads the contents, and checks the checksums.

This PR also contains a refactor of the Streamlets API to handle both byte and record oriented streamlets (but needs further testing). We have added a couple of sample Streamlets of each type for testing purposes. This PR is only the first preliminary version of the above changes. More testing is needed, especially with real streaming systems offloading data instead of just testing with S3 client.

---------

Signed-off-by: Raúl Gracia <raul.gracia@dell.com>
Co-authored-by: Raúl Gracia <raul.gracia@dell.com>